Added support for contracts 1.3.0
- Fixed tests
- bumped version to 0.9.2
