+++ contracts_lib_py/nft_upgradeable.py
@@ -140,3 +140,27 @@ class NFTUpgradeable(ContractBase):
         )
         return self.is_tx_successful(tx_hash)
 
+    def balance(self, address, did):
+        """
+        Returns the NFT balance for an address
+
+        :param address: Account address to check, str
+        :param did: The NFT id, str
+        :return: int
+        """
+        return self.contract.caller.balanceOf(address, int(did, 16))
+
+    def transfer_nft(self, did, address, amount, account):
+        tx_hash = self.send_transaction(
+            'safeTransferFrom',
+            (account.address,
+             address,
+             int(did, 16),
+             amount,
+             Web3.toBytes(text='')
+             ),
+            transact={'from': account.address,
+                      'passphrase': account.password,
+                      'keyfile': account.key_file}
+        )
+        return self.is_tx_successful(tx_hash)
