+++ setup.py
@@ -17,7 +17,7 @@ with open('CHANGELOG.md') as history_file:
 # Installed by pip install contracts-lib-py
 # or pip install -e .
 install_requirements = [
-    'nevermined-contracts==1.2.2',
+    'nevermined-contracts==1.3.0',
     'web3==5.13.0',
     'requests==2.21.0',
 ]
@@ -82,6 +82,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/nevermined-io/contracts-lib-py',
-    version='0.9.1',
+    version='0.9.2',
     zip_safe=False,
 )
